added condition checks

+++ index.js
@@ -115,6 +115,11 @@ mongoose.connect('mongodb://localhost:27017/myFlix', { useNewUrlParser: true, us
 
     // CREATE (POST) request - adding movie to favorites
         app.post('/users/:Username/movies/:MovieID', async (req, res) => {
+            // CONDITION TO CHECK ADDED HERE
+            if(req.user.Username !== req.params.Username){
+                return res.status(400).send('Permission denied');
+            }
+            // CONDITION ENDS
             await Users.findOneAndUpdate({ username: req.params.Username }, {
             $push: { favoriteMovies: req.params.MovieID }
             },
@@ -130,6 +135,11 @@ mongoose.connect('mongodb://localhost:27017/myFlix', { useNewUrlParser: true, us
 
     // DELETE (DELETE) request - delete movie from favorites
         app.delete('/users/:Username/movies/:MovieID', passport.authenticate('jwt', {session: false}), async (req, res) => {
+            // CONDITION TO CHECK ADDED HERE
+            if(req.user.Username !== req.params.Username){
+                return res.status(400).send('Permission denied');
+            }
+            // CONDITION ENDS
             await Users.findOneAndUpdate({ username: req.params.Username }, {
                 $pull: { FavoriteMovies: req.params.MovieID }
             },
@@ -145,6 +155,11 @@ mongoose.connect('mongodb://localhost:27017/myFlix', { useNewUrlParser: true, us
 
     // DELETE (DELETE) request - delete existing user
         app.delete('/users/:Username', passport.authenticate('jwt', {session: false}), async (req, res) => {
+            // CONDITION TO CHECK ADDED HERE
+            if(req.user.Username !== req.params.Username){
+                return res.status(400).send('Permission denied');
+            }
+            // CONDITION ENDS
             await Users.findOneAndDelete({ username: req.params.Username })
             .then((user) => {
                 if (!user) {
